un-hardcoded secret key, extracted user class to separate file from app.py
