+++ key.py
@@ -1,0 +1,5 @@
+import os
+
+print(os.urandom(32))
+
+# export SECRET_KEY="secretkey"
